Virtual Tourist App
An app that downloads and stores images from Flickr. The app allows
users to drop pins on a map, as if they were stops on a tour. Users are
then able to download pictures for the location and persist both the
pictures, and the association of the pictures with the pin.